Clarify sign/size of ABI types (#279)
Use u32/u64/usize everywhere (all ABI values are treated as unsigned),
with usize specifically indicating a linear memory address or size
(so any future Wasm64 port would map usize->u64 instead of u32).

+++ rust/oak/src/grpc/mod.rs
@@ -106,7 +106,7 @@ pub fn event_loop<T: OakNode>(
         crate::prep_handle_space(&mut space);
         unsafe {
             let status = wasm::wait_on_channels(space.as_mut_ptr(), read_handles.len() as u32);
-            match OakStatus::from_i32(status) {
+            match OakStatus::from_i32(status as i32) {
                 Some(OakStatus::OK) => (),
                 Some(err) => return err as i32,
                 None => return OakStatus::OAK_STATUS_UNSPECIFIED.value(),
